+++ CHANGELOG.md
@@ -5,6 +5,7 @@ To see version information, `xsvutils --version`.
 To see compatibility information, `xsvutils help compatibility`.
 
 ## version 0.2.5 (Unreleased)
+- fix a bug of removing temp files on `sort` subcommand
 
 ## version 0.2.4 (2018/03/30)
 - add support subquery of `tee` subcommand on v2