+++ CHANGELOG.md
@@ -6,6 +6,7 @@ To see compatibility information, `xsvutils help compatibility`.
 
 ## version 0.2.5 (Unreleased)
 - fix a bug of removing temp files on `sort` subcommand
+- fix a bug of output of `uriparams --name-list`
 
 ## version 0.2.4 (2018/03/30)
 - add support subquery of `tee` subcommand on v2